Fixing jekyll installation
Gem by default installs gems for the current user only - so if we do so as root, the 'app' user can't access the gem.

There is an override for this, which is now applied so the gem is available to everyone. However a second bug comes with the Jekyll binary being installed to '/usr/local/bin/' which is in the path of the root user on CentOS, but is on Ubuntu :/

To workaround this second bug, the tests in this role specify the binary path directly. I am not yet sure if this will cause problems in terms of running Jekyll within Ansible for example. If it is - we will likely need to make changes to the base images to 'fix' the PATH variable.