Implementing Pristine BARC flavour 0.2.0
Pristine is a base project designed to bootstrap new projects and give a consistent provisioning setup across all BAS projects.

A version of this base project has been created for BARC roles, which this commit implements for this Ruby2 role.

The majority of the changes made to this projects structure are a refactoring, rather than introducing new technologies or design methods.